fix: resolve elixir circular dependacy
