+++ packages/elixir/src/components/AddLiquidity/FeeSelector/index.tsx
@@ -2,8 +2,9 @@ import { Box, Button, Text } from '@honeycomb-finance/core';
 import { useChainId, usePrevious, useTranslation } from '@honeycomb-finance/shared';
 import { FeeAmount } from '@pangolindex/sdk';
 import React, { useCallback, useEffect, useMemo, useState } from 'react';
-import { useFeeTierDistributionHook, usePoolsHook } from 'src/hooks';
+import { usePoolsHook } from 'src/hooks';
 import { PoolState } from 'src/hooks/types';
+import { useFeeTierDistributionHook } from '../../../hooks/FeeTier';
 import FeeOption from './FeeOption';
 import { FeeTierPercentageBadge } from './FeeTierPercentageBadge';
 import { FEE_AMOUNT_DETAIL } from './shared';
