+++ packages/elixir/src/components/AddLiquidity/index.tsx
@@ -28,7 +28,8 @@ import LiquidityChartRangeInput from 'src/components/LiquidityChartRangeInput';
 import { Bound, Field } from 'src/state/mint/atom';
 import { useDerivedMintInfo, useMintActionHandlers, useMintState, useRangeHopCallbacks } from 'src/state/mint/hooks';
 import { useElixirAddLiquidityHook } from 'src/state/wallet/hooks';
-import { useDerivedPositionInfo, useElixirPositionFromTokenId } from 'src/state/wallet/hooks/evm';
+import { useDerivedPositionInfo } from 'src/state/wallet/hooks/evm';
+import { useElixirPositionFromTokenId } from 'src/state/wallet/hooks/useElixirPositionFromTokenId';
 import OutofRangeWarning from '../OutofRangeWarning';
 import ConfirmDrawer from './ConfirmDrawer';
 import FeeSelector from './FeeSelector';
